Split windows.rs into several files
